+++ .github/workflows/test.yml
@@ -22,7 +22,6 @@ jobs:
         cache: 'npm'
     - run: npm ci
     - run: npm test
-    - run: npm audit
-    - run: npm run lint
+    - run: npm run integration-test
       env:
         NODE_ENV: test
