+++ test/integration/index.spec.js
@@ -1,0 +1,1 @@
+import './repository/index.spec.js'
