+++ test/integration/repository/fetch/performance.spec.js
@@ -3,8 +3,8 @@ import assert from 'node:assert'
 import { test } from 'node:test'
 import ioredis from 'ioredis-mock'
 
-import { Repository } from '../../../src/repository.js'
-import { Chatroom } from '../../model/index.js'
+import { Repository } from '../../../../src/repository.js'
+import { Chatroom } from '../../../model/index.js'
 
 test('repository', async t => {
   await t.test('#fetch', async t => {
