+++ test/integration/repository/index.spec.js
@@ -1,0 +1,4 @@
+import './fetch/performance.spec.js'
+
+import './save/structure.spec.js'
+import './save/performance.spec.js'
