+++ test/integration/repository/save/structure.spec.js
@@ -2,8 +2,8 @@ import assert from 'node:assert'
 import { test } from 'node:test'
 import ioredis from 'ioredis-mock'
 
-import { Repository } from '../../../src/repository.js'
-import { Chatroom, Message } from '../../model/index.js'
+import { Repository } from '../../../../src/repository.js'
+import { Chatroom } from '../../../model/index.js'
 
 test('repository', async t => {
   let repo
