+++ test/repository/index.spec.js
@@ -1,7 +1,3 @@
 import './new.spec.js'
 import './fetch/fetch.spec.js'
-import './fetch/performance.spec.js'
-
 import './save/save.spec.js'
-import './save/structure.spec.js'
-import './save/performance.spec.js'
